Win32/MSVC: Fixed a compiler warning.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@20365 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/gettimeofday.c
@@ -28,6 +28,6 @@ gettimeofday(struct timeval *tv, struct timezone *tz)
 
     _ftime(&timebuffer);
     tv->tv_usec = timebuffer.millitm * 1000;
-    tv->tv_sec = timebuffer.time;
+    tv->tv_sec = (long)timebuffer.time;
     return (0);
 }
